build(description): add temporary dependency on openxlsx
The temporary support for Source Plate Contents requires xlsx parsing; so the library openxlsx is
temporarily added as a package dependency under Imports. The dependency chain for openxlsx is rather
bulky, so if for some reason xlsx parsing becomes a permanent feature of this library, openxlsx
should be removed and replaced with a similar library that doesn't pull Rcpp.

re #12