feat(r/primary_analysis.r): add interface and testing for dynamic well assignment
there seems to be a default 96->384 well mapping in use in the telomere analysis. as such, that well
assignment is hard coded as the default constructor of the PrimaryAnalysis S4 class. however, the
actual linker information is available if plate content and plate list are both specified, and it's
possible someone will want to change the default plate mapping someday. so this patch adds the
ability to specify the plate contents and list files to create.analysis, along with a flag to
activate the linker calculation, and input parameter consistency testing. the actual well mapping is
not yet implemented, and I don't really have the appropriate test cases to debug this for the
moment.

re #5 #12